# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KKN Team Website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: Get all announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get inventory list (list barang)
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /deadlines:
    get:
      operationId: getDeadlines
      tags: [deadlines]
      summary: Get deadlines list
      responses:
        "200":
          description: Deadlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deadline"
    post:
      operationId: createDeadline
      tags: [deadlines]
      summary: Create deadline (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeadlineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"
  /deadlines/{id}:
    patch:
      operationId: updateDeadline
      tags: [deadlines]
      summary: Update deadline (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeadlineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"
    delete:
      operationId: deleteDeadline
      tags: [deadlines]
      summary: Delete deadline (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /complaints:
    get:
      operationId: getComplaints
      tags: [complaints]
      summary: Get daily life complaints (keluhan)
      responses:
        "200":
          description: Complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create complaint (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    delete:
      operationId: deleteComplaint
      tags: [complaints]
      summary: Delete complaint (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /issues:
    get:
      operationId: getIssues
      tags: [issues]
      summary: Get KKN program issues (masalah proker)
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create issue (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}:
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update issue (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete issue (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /kas:
    get:
      operationId: getKas
      tags: [kas]
      summary: Get all kas transactions
      responses:
        "200":
          description: Kas list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kas"
    post:
      operationId: createKas
      tags: [kas]
      summary: Create kas entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KasInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kas"
  /kas/{id}:
    patch:
      operationId: updateKas
      tags: [kas]
      summary: Update kas entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KasUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kas"
    delete:
      operationId: deleteKas
      tags: [kas]
      summary: Delete kas entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Announcement:
      type: object
      required: [id, title, content, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content, priority]
      properties:
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
...
    InventoryItem:
      type: object
      required: [id, name, category, quantity, unit, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [p3k, obat, alkes, umum]
        quantity:
          type: integer
        unit:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InventoryItemInput:
      type: object
      required: [name, category, quantity, unit]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [p3k, obat, alkes, umum]
        quantity:
          type: integer
        unit:
          type: string
        notes:
          type: string
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [p3k, obat, alkes, umum]
        quantity:
          type: integer
        unit:
          type: string
        notes:
          type: string
    Deadline:
      type: object
      required: [id, title, type, dueDate, status, assignedTo, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [tugas, kegiatan]
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, done]
        assignedTo:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DeadlineInput:
      type: object
      required: [title, type, dueDate, status, assignedTo]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [tugas, kegiatan]
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, done]
        assignedTo:
          type: array
          items:
            type: string
        notes:
          type: string
    DeadlineUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [tugas, kegiatan]
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, done]
        assignedTo:
          type: array
          items:
            type: string
        notes:
          type: string
    Complaint:
      type: object
      required: [id, title, description, reportedBy, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reportedBy:
          type: string
        status:
          type: string
          enum: [open, resolved]
        createdAt:
          type: string
    ComplaintInput:
      type: object
      required: [title, description, reportedBy, status]
      properties:
        title:
          type: string
        description:
          type: string
        reportedBy:
          type: string
        status:
          type: string
          enum: [open, resolved]
    ComplaintUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        reportedBy:
          type: string
        status:
          type: string
          enum: [open, resolved]
    Issue:
      type: object
      required: [id, title, description, category, priority, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [proker, administrasi, logistik, lainnya]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [open, in_progress, resolved]
        createdAt:
          type: string
    IssueInput:
      type: object
      required: [title, description, category, priority, status]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [proker, administrasi, logistik, lainnya]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [open, in_progress, resolved]
    IssueUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [proker, administrasi, logistik, lainnya]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [open, in_progress, resolved]
...
    Kas:
      type: object
      required: [id, type, amount, description, category, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [pemasukan, pengeluaran]
        amount:
          type: integer
        description:
          type: string
        category:
          type: string
          enum: [makan, transport, perlengkapan, administrasi, lainnya]
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    KasInput:
      type: object
      required: [type, amount, description, category, date]
      properties:
        type:
          type: string
          enum: [pemasukan, pengeluaran]
        amount:
          type: integer
        description:
          type: string
        category:
          type: string
          enum: [makan, transport, perlengkapan, administrasi, lainnya]
        date:
          type: string
        notes:
          type: string
    KasUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [pemasukan, pengeluaran]
        amount:
          type: integer
        description:
          type: string
        category:
          type: string
          enum: [makan, transport, perlengkapan, administrasi, lainnya]
        date:
          type: string
        notes:
          type: string
...